requirements.txt
streamlit
openai
pandas
aiohttp
scikit-learn
fastapi>=0.110
uvicorn[standard]>=0.24
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31
